requirements.txt
aioredis==1.3.1; python_version >= "3.7"
anyio==3.5.0; python_version >= "3.7" and python_full_version >= "3.6.2"
argon2-cffi-bindings==21.2.0; python_version >= "3.8"
argon2-cffi==21.3.0; python_version >= "3.8"
asgiref==3.5.0; python_version >= "3.8"
async-timeout==4.0.2; python_version >= "3.7"
attrs==21.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
autobahn==22.3.2; python_version >= "3.7"
automat==20.2.0; python_full_version >= "3.6.7" and python_version >= "3.6"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
bcrypt==3.2.0; python_version >= "3.8"
better-exceptions==0.3.3
brotli==1.0.9; python_version >= "3.7"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0
channels-redis==3.4.0; python_version >= "3.7"
channels==3.0.4; python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
click==8.1.2; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
constantly==15.1.0; python_full_version >= "3.6.7" and python_version >= "3.6"
cryptography==36.0.2
daphne==3.0.2; python_version >= "3.7"
dealer==2.1.0
defusedxml==0.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
dj-database-url @ git+https://github.com/jacobian/dj-database-url.git@master
django-allauth==0.50.0
django-appconf==1.0.5; python_version >= "3.6"
django-compressor==3.1
django-dbbackup @ git+https://github.com/django-dbbackup/django-dbbackup.git@master ; python_version >= "3.6"
django-debug-toolbar==3.2.4; python_version >= "3.6"
django-ranged-response==0.2.0
django-sendgrid-v5==1.2.0
django-simple-captcha==0.5.17
django==4.0.4; python_version >= "3.8"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.13.0; python_version >= "3.7"
h2==3.2.0; python_full_version >= "3.6.7" and python_version >= "3.6"
hiredis==2.0.0; python_version >= "3.7"
hpack==3.0.0; python_full_version >= "3.6.7"
httptools==0.4.0; python_version >= "3.7" and python_full_version >= "3.5.0"
hyperframe==5.2.0; python_full_version >= "3.6.7"
hyperlink==21.0.0; python_full_version >= "3.6.7" and python_version >= "3.7"
idna==3.3
incremental==21.3.0; python_full_version >= "3.6.7" and python_version >= "3.6"
msgpack==1.0.3; python_version >= "3.7"
numpy==1.22.3; python_version >= "3.8"
oauthlib==3.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pillow==9.1.0; python_version >= "3.7"
prettytable==2.5.0; python_version >= "3.6"
priority==1.3.0; python_full_version >= "3.6.7" and python_version >= "3.6"
psycopg2==2.9.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_full_version >= "3.6.7"
pyasn1==0.4.8; python_full_version >= "3.6.7"
pycparser==2.21; python_version >= "3.8" and python_full_version < "3.0.0" or python_version >= "3.8" and python_full_version >= "3.4.0"
pyjwt==2.3.0; python_version >= "3.6"
pymemcache==3.5.2
pyopenssl==22.0.0; python_full_version >= "3.6.7" and python_version >= "3.6"
python-dotenv==0.20.0; python_version >= "3.7"
python-http-client==3.3.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
python3-openid==3.2.0
pytz==2022.1; python_version >= "3.6"
pyyaml==6.0; python_version >= "3.7"
rcssmin==1.1.0
requests-oauthlib==1.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.27.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
rjsmin==1.2.0
scipy==1.6.1; python_version >= "3.7"
sendgrid==6.9.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
service-identity==21.1.0; python_full_version >= "3.6.7" and python_version >= "3.6"
six==1.16.0
sniffio==1.2.0; python_version >= "3.7" and python_full_version >= "3.6.2"
sqlparse==0.4.2; python_version >= "3.8"
starkbank-ecdsa==2.0.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
twisted-iocpsupport==1.0.2; platform_system == "Windows" and python_full_version >= "3.6.7" and python_version >= "3.6"
twisted==21.7.0; python_full_version >= "3.6.7"
txaio==22.2.1; python_version >= "3.7"
typing-extensions==4.1.1; python_version >= "3.6" and python_full_version >= "3.6.7"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
uvicorn==0.17.6; python_version >= "3.7"
uvloop==0.16.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.8.2; python_version >= "3.7"
wcwidth==0.2.5; python_version >= "3.6"
websockets==10.2; python_version >= "3.7"
whitenoise==6.0.0; python_version >= "3.7"
zope.interface==5.4.0; python_full_version >= "3.6.7" and python_version >= "3.6"
